Split C++ implementation too
