+++ rust/crates/eckit-sys/build.rs
@@ -41,15 +41,23 @@ fn build_cxx_bridge(include: &std::path::Path) {
     let out_dir = std::path::PathBuf::from(std::env::var("OUT_DIR").expect("OUT_DIR not set"));
 
     println!("cargo:rerun-if-changed=cpp/eckit_bridge.h");
-    println!("cargo:rerun-if-changed=cpp/eckit_bridge.cpp");
     println!("cargo:rerun-if-changed=cpp/log.h");
+    println!("cargo:rerun-if-changed=cpp/log.cpp");
     println!("cargo:rerun-if-changed=cpp/config.h");
+    println!("cargo:rerun-if-changed=cpp/config.cpp");
     println!("cargo:rerun-if-changed=cpp/datahandle.h");
+    println!("cargo:rerun-if-changed=cpp/datahandle.cpp");
     println!("cargo:rerun-if-changed=cpp/message.h");
+    println!("cargo:rerun-if-changed=cpp/message.cpp");
     println!("cargo:rerun-if-changed=cpp/stream.h");
+    println!("cargo:rerun-if-changed=cpp/stream.cpp");
 
     cxx_build::bridge("src/lib.rs")
-        .file(crate_dir.join("cpp/eckit_bridge.cpp"))
+        .file(crate_dir.join("cpp/log.cpp"))
+        .file(crate_dir.join("cpp/config.cpp"))
+        .file(crate_dir.join("cpp/datahandle.cpp"))
+        .file(crate_dir.join("cpp/message.cpp"))
+        .file(crate_dir.join("cpp/stream.cpp"))
         .include(include)
         .include(crate_dir.join("cpp"))
         .include(&out_dir) // for eckit_exceptions.h
